% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lifehist.R
\name{CalcFirstLastRepro}
\alias{CalcFirstLastRepro}
\title{Calculate potential reproductive years}
\usage{
CalcFirstLastRepro(
  LHextra,
  AgeFirstRepro = c(1, 1),
  AgeLastRepro = c(100, 100),
  YearLastRepro = c(0, 0),
  MonthLastRepro = c(1, 1)
)
}
\arguments{
\item{LHextra}{datAgeFirstReproame with lifehistory data, with columns id - Sex -
BirthYear - DeathYear - DeathMonth. Missing values allowed.}

\item{AgeFirstRepro}{vector of length 2, population minimum age of first
reproduction for females and males, respectively.}

\item{AgeLastRepro}{vector of length 2, population maximum age of last
reproduction for females and males, respectively.}

\item{YearLastRepro}{vector of length 2, maximum calendar years of last
reproduction after calendar year of death.}

\item{MonthLastRepro}{vector of length 2, minimum calendar month of last
reproduction in calendar year of death. Note for females that foetusses
may be sampled.}
}
\value{
\code{LHextra} with 2 added columns:
\describe{
   \item{FirstRepro}{The (minimum) year in which the individual could first
   have reproduced. NA for individuals that died before the age of first
   reproduction.}
   \item{LastRepro}{The (maximum) year in which the individual could last
   have reproduced}}
}
\description{
Calculate minimum year of first reproduction and maximum year of
  last reproduction, based on individual birth year, death year and month,
  and population age of first and last reproduction.
}
\details{
If the breeding season is in September to November and offspring are
  born the next spring, than a male which died in September of year \emph{t} may
  still have sired offspring born in year \emph{t+1}: \code{YearLastRepro = 1} and
  \code{MonthLastRepro = 9}. A male which died in August cannot be the sire
  of any offspring born in year \emph{t+1}.
}
\examples{
\dontrun{
LHextra <- CalcFirstLastRepro(LHextra,
                             AgeFirstRepro = c(3, 2),
                             AgeLastRepro = c(25, 20),
                             MonthLastRepro = c(1, 9),
                             YearLastRepro = c(0, 1))
LH.chk <- CheckBirthYears(Pedigree,
                         LHextra,
                         AgeFirstRepro = c(3, 2),
                         AgeLastRepro = c(25,20))
}

}
\seealso{
\code{\link{CheckBirthYears}}
}
