% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lifehist.R
\name{CalcParAge}
\alias{CalcParAge}
\title{Calculate parental age}
\usage{
CalcParAge(
  candpar.df,
  LHextra,
  Factor = TRUE,
  BRKS = c(0, 4, 5, 6, 11, 12, 14, 25),
  LBLS = c(3.5, 5, 6, 9, 12, 13.5, 15),
  REFLVL = "9",
  Dead = -42
)
}
\arguments{
\item{candpar.df}{dataframe with columns id - par.ped - par.field}

\item{LHextra}{dataframe with (at least) columns id - BirthYear - LastRepro}

\item{Factor}{logical, return ages as a factor using the specified
breakpoints (TRUE) or as a numeric vector (FALSE)}

\item{BRKS}{breaks to pass to \code{\link{cut}}}

\item{LBLS}{labels to pass to \code{\link{cut}}}

\item{REFLVL}{reference level for factor}

\item{Dead}{value in case the id's birth year is later than the field
parent's last year of reproduction}
}
\value{
a numeric vector or factor of length equal to number of rows in
  candpar.df, and in that exact same order.
}
\description{
wrapper to calculate age of candidate field parent
}
\details{
the default factor levels are Rum red deer specific
}
\examples{
\dontrun{
cand.sires <- Init.candpar(unique(Pedigree[,c("id", "sire")]), FieldSires,
			                             expl.var=colnames(FieldSires)[3])
cand.sires$Age.f <- CalcParAge(cand.sires, LHextra, Factor=TRUE)
cand.sires$prob <- predict(DadMod, newdata=cand.sires, type="response")
}

}
