% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lifehist.R
\name{CheckBirthYears}
\alias{CheckBirthYears}
\title{Check consistency of birth years and potential reproductive years
  throughout pedigree}
\usage{
CheckBirthYears(
  Pedigree,
  LHextra,
  AgeFirstRepro = c(1, 1),
  AgeLastRepro = c(100, 100)
)
}
\arguments{
\item{Pedigree}{dataframe with columns id - dam - sire}

\item{LHextra}{datframe with lifehistory data, with columns id - Sex -
BirthYear - FirstRepro - LastRepro.}

\item{AgeFirstRepro}{vector of length 2, population minimum age of first
reproduction for females and males, respectively.}

\item{AgeLastRepro}{vector of length 2, population maximum age of last
reproduction for females and males, respectively.}
}
\value{
a dataframe with the following columns
  \item{id}{}
  \item{dam}{}
  \item{sire}{}
  \item{Sex}{1=females,2=males,3=unknown. Pedigree dam/sire takes precedent
    over LHextra}
  \item{BirthYear.self, FirstRepro.self, LastRepro.self}{based on own data in
    LHextra}
  \item{BY.min.dam, BY.max.dam}{based on dam FirstRepro & LastRepro}
  \item{BY.min.sire, BY.max.sire}{based on sire FirstRepro & LastRepro}
  \item{FirstRepro.off, LastRepro.off}{First and last birth year of pedigree
    offspring}
  \item{BY.min.off, BY.max.off}{based on offspring birth years & AgeFirstRepro}
  \item{OK.dam, OK.sire, OK.off}{dam, sire, offspring lifehistory data
    consistent with own data (logical)}
  \item{OK.dam.off, OK.sire.off}{dam, sire data consistent with offspring
    lifehistory data (logical)}
  \item{BY.min, BY.max}{Consensus minimum & maximum birth year}
  \item{OK}{everything OK/not (logical)}
}
\description{
Check that offspring birth years are consistent with their
  assigned parents' potential reproductive years as calculated by
  \code{\link{CalcFirstLastRepro}} based on their birth year, death year and
  month, and population age of first and last reproduction.
}
\examples{
\dontrun{
LHextra <- CalcFirstLastRepro(LHextra,
                             AgeFirstRepro = c(3, 2),
                             AgeLastRepro = c(25, 20),
                             MonthLastRepro = c(1, 9),
                             YearLastRepro = c(0, 1))
LH.chk <- CheckBirthYears(Pedigree,
                         LHextra,
                         AgeFirstRepro = c(3, 2),
                         AgeLastRepro = c(25,20))
}

}
