% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RumDeer_DaysHeld_to_ProbSire.R
\name{DaysHeld_to_Prob}
\alias{DaysHeld_to_Prob}
\title{Days held during conception window to sire probability}
\usage{
DaysHeld_to_Prob(ObsSire = NULL, Model = "default", Plot = TRUE)
}
\arguments{
\item{ObsSire}{dataframe with the following columns:
\describe{
  \item{id}{}
  \item{par.ped}{genetically inferred parent; 1 per id, or NA)}
  \item{par.field}{Field observed candidate parent; 0, 1 or several per id}
  \item{Ndays}{number of days \code{id}'s mother was in \code{par.field}'s harem during
    the 11-day conception window.}
  \item{age.par.field}{age in years of \code{par.field} at \code{id}'s birth. Only
    used when \code{Model = 'age'}} }}

\item{Model}{binomial glm to calculate the probability that a male is the
father, using column names in DaysHeld. Implemented are
\describe{
  \item{\code{'default'}}{fits Ndays and Ndays^2}
  \item{\code{'age'}}{additionally fits age as a factor with levels 2-4, 5,
6, 7-11, 12, 13-14, 15+.} }}

\item{Plot}{plot observed & predicted values?}
}
\value{
the fitted glm model
}
\description{
Calculate the probability the male is the offspring's sire from
  the number of days its mother was in his harem during the 11-day window
  around back-calculated conception (birth date - 235 days).
}
\details{
The sum of Ndays per offspring may exceed 11, as Some females are
  recorded with 2 different males on the same day. Calculating part-days for
  these cases was tried, but does not notably affect the relationship between
  days held and probability to sire the offspring.

  The age levels were determined by fitting a model with each age as a single
  level, and iteratively combining ages which effects on paternity
  probability did not differ significantly.
}
\section{Database query}{

The in-build query "Sys_BehaviouralPaternities" returns candidate fathers in
whose harem the mother was seen at least 6 days during the conception window.
To obtain all candidate fathers, including those in whose harem the mother
was seen 1-5 days, build a 'Totals' query around "sys_ConceptionBirthDates"
and "tblRutcharts", with
\itemize{
  \item sys_ConceptionBirthDates.Code (Group By)
  \item sys_ConceptionBirthDates.MumCode linked to tblRutCharts.HindCode (Group By)
  \item tblRutCharts.Date (Where; do not show)
  \item Criteria = Abs([sys_ConceptionBirthDates.ConceptPredict]-[tblRutCharts.Date])<6
  \item tblRutCharts.StagCode (Group By)
  \item Criteria = Not "NONE"
  \item Ndays: tblRutCharts.Date (Count) }
}

\examples{
\dontrun{
candsires.tmp <- Init.candpar(Pedigree[,c("id", "sire")], FieldCandSires, expl.var="Ndays")
candsires.tmp$age.par.field <- CalcParAge(candsires.tmp, LHextra, Factor=FALSE)
candsires.tmp$Age.f <- CalcParAge(candsires.tmp, LHextra, Factor=TRUE, REFLVL="9")

# subset those to run model on. don't use Ndays=0: not all present/alive that
# year. Suggest to use only those where both id and par.field are SNP genotyped,
# such that it is known with high accuracy whether or not par.field == par.ped.
# par.ped may be NA, which is treated as par.field != par.ped.
candsires.tmp$Use <- with(candsires.tmp, ifelse(Ndays == 0, FALSE,
                                     ifelse(is.na(age.par.field), FALSE,
                                        ifelse(par.field \%in\% Pedigree$id
                                               & id \%in\% Pedigree$id, TRUE, FALSE))))
candsires.tmp <- candsires.tmp[candsires.tmp$Use, ]

mod.DaysHeld <- DaysHeld_to_Prob(ObsSire = candsires.tmp, Model = "age", Plot=TRUE)
}

}
