% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offspring.R
\name{Init.candpar}
\alias{Init.candpar}
\title{Initiate the dataframe with candidate parents}
\usage{
Init.candpar(ParPedigree, ParField, expl.var = "prob")
}
\arguments{
\item{ParPedigree}{parent in genetically inferred pedigree, dataframe with
columns id-dam or id-sire, column names ignored and renamed to 'id' and
'par.ped'. May include dummy individuals and mystery samples.}

\item{ParField}{candidate parents inferred in the field, dataframe with
columns id - par.field - [expl.var].}

\item{expl.var}{explanatory variable to calculate probability that
  par.field is identical to par.ped. Must be a column name in `ParField'.

   Defaults to 'prob', and id-par.field
  combinations not listed in 'ParField', as well as 'INVIS', will become
  'NA'.}
}
\value{
a dataframe with columns id - par.ped - par.field - [expl.var]
}
\description{
For all individuals sharing a pedigree parent, all candidate
  field parents of at least 1 member are listed, as well as 'INVIS'.
}
\details{
Id's sharing the same par.ped are considered a sibship cluster, and
  all par.field occuring with at least one id in the cluster are considered a
  candidate parent for the whole cluster. Resulting Id - par.field pairs not
  occuring in ParField are given 'NA' for expl.var, except when
  expl.var="Ndays", when they get '0'.

  If \code{expl.var='carer'}, there will be a warning about pedigree sibling
  clusters with >1 carer.

  'INVIS' is added as candidate parent to all sibship clusters. If
  \code{expl.var='Ndays'}, the expl.var value for 'INVIS' for each id is
  calculated as 11 - the sum of Ndays with known par.field for that id,
  otherwise expl.var = NA for INVIS.
}
\examples{
\dontrun{
FieldMums <- setNames(tblLife[!is.na(tblLife$MumCode), c("Code", "MumCode")],
                     c("id", "par.field"))
FieldMums$carer <- 1
cand.dams <- Init.candpar(unique(Pedigree[,c("id", "dam")]), FieldMums,
                         expl.var="carer")

cand.sires <- Init.candpar(Pedigree[,c("id", "sire")],
                          setNames(DaysHeld[, c("Code","StagCode","Ndays")],
                                   c("id", "par.field", "Ndays")),
                          expl.var="Ndays")
}

}
\seealso{
\code{\link{parDFtoM}, \link{ParentOfAll}}
}
