% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.R
\name{init.PedFieldMatch}
\alias{init.PedFieldMatch}
\title{Initiate array 'PedFieldMatch'}
\usage{
init.PedFieldMatch()
}
\value{
a 5D-array is created globally, taking the following values:
  \item{OK}{match}
  \item{X}{not a match}
  \item{LIKELY}{Conclusive based on offspring, not SNPd, but undetermined if
  dummy mother matches}
  \item{MAYBE}{maternal sibling or match: Mother matches, not SNPd,
  no/inconclusive based on offspring, Sex & Years match/unknown, }
  \item{MERGE?}{one id plausible based on label, and another based on
  offspring}
  \item{CHECKMUM}{id plausible based on label/offspring, but mum in pedigree
  & not in field, or mismatches}
  \item{CHECKYEARS}{id plausible based on label/offspring, but inconsistency
  in birth/reproductive years}
  \item{CHECKSEX}{id plausible based on label/offspring, but sex mismatch}
  \item{???}{Not enough info to assess plausibility}

Dimensions 2-5 are as returned by \code{\link{MkIDcat}}:
\item{Label}{Match - MismatchSNPd - MismatchNotSNPd - NotSNPd; 'SNPd' and
  'NotSNPd' refers to the field ID}
\item{Sex}{Match - Mismatch - Unknown}
\item{Years}{Match - Mismatch - Unknown}
\item{Mother}{Match - Mismatch - NewMum (only pedigree dam) - Unverified
(only field dam) - Unknown (dummy or mystery dam in pedigree) - None}
\item{Offspring}{Conclusive (is best match) - Inconclusive - Inferior (different candidate is best match) - Invisible ('INVIS' is best match) - None}
}
\description{
This array is used to 'translate' the various levels in the
  array returned by \code{\link{MkIDcat}}, plus offspring-based likelihoods
  being conclusive/inconclusive or none, for no offspring, into a consensus.
}
\details{
For the Offspring-based category, 'Invisible' denotes that an
  unobserved individual is the most likely match. Note that this may be one
  of the (other, shortlisted) candidate field parents, if they were not
  observed during the relevant breeding period(s).
}
\seealso{
\code{\link{MatchPedField}}
}
