% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchIDsPedigreeField.R
\name{MatchIDsPedigreeField}
\alias{MatchIDsPedigreeField}
\title{Match IDs between a genetic pedigree and field data}
\usage{
MatchIDsPedigreeField(
  Pedigree = NULL,
  LHextra = NULL,
  FieldDams = NULL,
  FieldSires = NULL,
  FieldIDs = NULL,
  MumProbs = c(CarerIsMum = 0.995, NoCarer_IsMum = 0.05),
  DadMod = NULL,
  AgeFirstRepro = c(3, 2),
  AgeLastRepro = c(25, 20),
  YearLastRepro = c(0, 1),
  MonthLastRepro = c(9, 1),
  nIter = 5,
  nInvis = c(3, 7),
  Lref = NULL,
  T.offcat = c(5, 3, 2, rep(1, nIter - 2)),
  nag = TRUE
)
}
\arguments{
\item{Pedigree}{genetically inferred pedigree, dataframe with columns
id-dam-sire. May include dummy individuals and mystery samples.}

\item{LHextra}{dataframe with lifehistory data, with columns id - Sex -
BirthYear - FirstRepro - LastRepro. See \code{\link{CalcFirstLastRepro}}.
May include field IDs (genotyped & not-genotyped) as well as mystery
samples.}

\item{FieldDams}{dataframe with id - dam.field - [expl.var], where the 3rd
column is a variable associated with the probability of parentage (e.g.
'carer')}

\item{FieldSires}{dataframe with id - sire.field - [expl.var], as
'FieldDams' (with expl.var e.g. 'Ndays')}

\item{FieldIDs}{character vector with IDs of real individuals (not mystery
samples), both genotyped and non-genotyped. If \code{NULL}, defaults to all
id's in LHextra, and all id's and field parents in FieldDams and
FieldSires.}

\item{MumProbs}{numeric vector length 2, with probability that field mum
(carer) is the pedigree mum, and probability that an arbitrary female is
the pedigree mum for an individual without field mum. [Note: having a GLM
model as input as for sires is fairly straightforward to implement.]}

\item{DadMod}{A fitted binomial GLM, with pedigree parent/not as response
variable and [expl.var] in FieldSires and optionally field parent age as
explanatory variable.}

\item{AgeFirstRepro}{vector of length 2, population minimum age of first
reproduction for females and males, respectively.}

\item{AgeLastRepro}{vector of length 2, population maximum age of last
reproduction for females and males, respectively.}

\item{YearLastRepro}{vector of length 2, maximum calendar years of last
reproduction after calendar year of death.}

\item{MonthLastRepro}{vector of length 2, minimum calendar month of last
reproduction in calendar year of death; see
\code{\link{CalcFirstLastRepro}} for details.}

\item{nIter}{Number of iterations}

\item{nInvis}{vector of length 2, number of 'invisible' candidate parents
per pedigree sibship; see \code{\link{ParentOfAll}}.}

\item{Lref}{list with 2 numeric vectors, 'dam' and 'sire', with reference
log10-probabilities for candidate parents as a function of sibship size;
the probability of a candidate field parent must at least pass this value,
see \code{\link{MkOffCat}}.}

\item{T.offcat}{Numeric vector of length 'nIter'; threshold log10-probability
between most-likely and next-most-likely candidate parent to declare a
conclusive match, see \code{\link{MkOffCat}}}

\item{nag}{logical, wait for user confirmation to continue when encountering
mismatches (TRUE), or always continue with fingers crossed (FALSE)?}
}
\value{
A list with
\describe{
  \item{Pedigree}{The pedigree, with id's replaced by field id's where
  possible, including in dam and sire columns. Rownames show the original id,
  and columns 'type.id', 'type.dam', and 'type.sire' indicate whether the
  name in the respective column is a proper, \strong{G}enotyped pedigree +
  field ID ('G'); or is a dummy/mystery pedigree ID \strong{R}eplaced by a
  field ID ('R'); or is a non-replaced \strong{D}ummy/mystery pedigree ID
  ('D')}
  \item{Likely.r}{a list of length nIter with dataframes with shortlisted
    candidate field ids for each pedigree id}
  \item{Match.r}{a list of length nIter with dataframes with accepted
    matches (a subset of the corresponding Likely.r element)} }

When \code{nag=TRUE} and choosing not to continue when a mismatch is
encountered, the object \code{Match.dump} is created in the Global
environment, which is a list with a selected subset of intermediate results
}
\description{
Match dummy and mystery IDs in the pedigree to field IDs, and
  check for consistency between pedigree and field-identified (candidate)
  parents.
}
\details{
All matches i, j with consensus level 'OK' are accepted, as well
  as 'LIKELY' in the second and third iteration, and 'CHECKMUM',
  'CHECKYEARS', and 'CHECKSEX' in subsequent iterations. Warnings are issued
  about 'MISMATCH', 'MERGE?', and 'CHECKMUM', as well as of cases where one
  field ID is accepted for two or more pedigree IDs (receiving consensus
  level 'MULTI'). For explanation of these levels, see
  \code{\link{init.PedFieldMatch}}
}
\examples{
\dontrun{
OUT <- MatchIDsPedigreeField(Pedigree = Pedigree,
                             LHextra = LHextra,
                             FieldDams = FieldMums,
                             FieldSires = FieldCandSires,
                             FieldIDs = FieldIDs,
                             MumProbs = c("CarerIsMum" = 0.995,
                                          "NoCarer_IsMum" = 0.05),
                             DadMod = mod.DaysHeld,
#                             Repro = list(
                                 AgeFirstRepro = c(3, 2),
                                 AgeLastRepro = c(25,20),
                                 MonthLastRepro = c(9,1),
                                 YearLastRepro = c(0, 1), #)
                             nInvis = c(3, 7),
                             Lref = list(dam = Lref.dam,
                                         sire = Lref.sire),
                             nIter = 5,
                             nag=TRUE)
}

}
