% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.R
\name{MatchPedField}
\alias{MatchPedField}
\title{Consensus match}
\usage{
MatchPedField(Offspring, Cat.i)
}
\arguments{
\item{Offspring}{vector with categories describing the match based on
pedigree offspring; output of \code{\link{MkOffCat}}}

\item{Cat.i}{matrix with 4 columns (Label - Sex - Years - Mother) and one row
per shortlisted candidate field id (optionally including INVIS); subset of
output of \code{\link{MkIDcat}}}
}
\value{
a vector with length equal to the number of field ids (number of rows
  in 'DF'), taking values
  \item{OK}{match}
  \item{X}{not a match}
  \item{LIKELY}{Conclusive based on offspring, not SNPd, but undetermined if
  dummy mother matches}
  \item{MAYBE}{maternal sibling or match: Mother matches, not SNPd,
  no/inconclusive based on offspring, Sex & Years match/unknown, }
  \item{MERGE?}{one id plausible based on label, and another based on
  offspring}
  \item{CHECKMUM}{id plausible based on label/offspring, but mum in pedigree
  & not in field, or mismatches}
  \item{CHECKYEARS}{id plausible based on label/offspring, but inconsistency
  in birth/reproductive years}
  \item{CHECKSEX}{id plausible based on label/offspring, but sex mismatch}
  \item{???}{Not enough info to assess plausibility}
}
\description{
Consensus category if the pedigree id matches any of the
  shortlisted candidate field ids, based on label, sex, birth years, pedigree
  mother's field offspring and pedigree offspring's field candidate parents.
}
\examples{
\dontrun{
Likely.L <- setNames(vector("list", length=dim(Cat.ID)[2]), unique(Pedigree$id))
for (i in 1:dim(Cat.ID)[2]) {
  these <- which(Cat.ID["Label",i,] == "Match" |
                 Cat.ID["Mother",i, ] == "Match" |
                 !is.na(LL.off[i,]))
  if (length(these)==0)  next
  Cat.off <- MkOffCat(LL.off[i,][these, drop=FALSE],  REF=Ref.i[i], T.off=3)
  Cat.i <- t(Cat.ID[,i,][,these,drop=FALSE])
  Consensus <- MatchPedField(Offspring=Cat.off, Cat.i)
  Likely.L[[i]] <- data.frame("Cand.id" = rownames(Cat.i),
                              Cat.i,
                              Off = Cat.off,
                              L.off = LL.off[i, these],
                              LR.off = LL.off[i,these] - Ref.i[i],
                              OK = Consensus,
                              stringsAsFactors = FALSE)
}
}

}
\seealso{
\code{\link{MkIDcat}, \link{MkOffCat}, \link{init.PedFieldMatch}}
}
