% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.R
\name{MkIDcat}
\alias{MkIDcat}
\title{Are pedigree and field ID synonymous?}
\usage{
MkIDcat(LH.pedigree = NULL, LH.field = NULL, ids.snpd.field = NULL)
}
\arguments{
\item{LH.pedigree}{genetically inferred pedigree, and 'pedigree-assisted'
Sex, BY.min (minimum birth year), BY.max (maximum birth year), see
\code{\link{CalcFirstLastRepro}}. May include dummy individuals and mystery
samples.}

\item{LH.field}{dataframe with lifehistory data based solely on field
observations, with columns id - dam.field - Sex - BY.min - BY.max. Must
have one row for each 'FieldID' (i.e. non-mystery, non-dummy IDs), even if
there is no associated lifehistory data known.}

\item{ids.snpd.field}{character vector with IDs that are both SNP genotyped
and proper field IDs; necessary after some dummy/mystery IDs in pedigree
have been replaced.}
}
\value{
An categorical array with 3 dimensions: Label-Sex-Years-
  Mother by no. individuals in LH.pedigree by the no. individuals in
  LH.field +1 ('INVIS'). The following categories are distinguished:
\item{Label}{Match - MismatchSNPd - MismatchNotSNPd - NotSNPd}
\item{Sex}{Match - Mismatch - Unknown (either one unknown sex)}
\item{Years}{Match - Mismatch - Unknown}
\item{Mother}{Match - Mismatch - NewMum (only pedigree dam) - Unverified
(only field dam) - Unknown (dummy or mystery dam in pedigree) - None}
}
\description{
Match IDs in the pedigree, including dummy and mystery IDs, to
  field IDs based on sample label, sex, birth year and reproductive years,
  and field-observed offspring of pedigree mother.
}
\examples{
\dontrun{
LH.chk <- CheckBirthYears(Pedigree, LHextra, AgeFirstRepro = AgeFirstRepro,
                          AgeLastRepro = AgeLastRepro)
LH.pedigree <- LH.chk[match(Pedigree$id, LH.chk$id), ]
LH.field <- merge(LHextra, FieldDams, all=TRUE)
LH.field <- LH.field[match(FieldIDs, LH.field$id), ]
Cat.ID <- MkIDcat(LH.pedigree, LH.field,
                  ids.snpd.field = intersect(rownames(pedigree), FieldIDs))
}

}
\seealso{
\code{\link{MkOffCat}, \link{MatchPedField}}
}
