% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match.R
\name{MkOffCat}
\alias{MkOffCat}
\title{Offspring-based categories}
\usage{
MkOffCat(LLO, REF, T.off = 3)
}
\arguments{
\item{LLO}{named numerical vector for a single pedigree ID with
log10-probabilities to match a subset of field IDs}

\item{REF}{reference log10-probability, minimum probability required for
'Conclusive'.}

\item{T.off}{threshold log10-probability difference between most-likely and
next-most-likely candidate. see 'Value'.}
}
\value{
A character vector of the same length as 'LLO', with the following
  possible values:
  \item{None}{No offspring in the pedigree}
  \item{Conclusive}{This field id is more likely than the next best
    by a margin of at least 'T.off', and is at least 'REF'.}
  \item{Inferior}{A different field id is more likely, by a margin of at
  least 'T.off'}
 \item{Inconclusive}{The likelihoods of the best and next-best candidate
 differ by less than 'T.off', or the most-likely is less than 'REF'}
 \item{Invisible}{'INVIS' is conclusively the most likely candidate}
}
\description{
Turn pedigree vs. field parent probabilities into categorical
  variables
}
\examples{
\dontrun{
for (i in 1:dim(Cat.ID)[2]) {
  these <- which(Cat.ID["Label",i,] == "Match" |
                 Cat.ID["Mother",i, ] == "Match" | !is.na(LL.off[i,]))
 if (length(these)>0) {
   Cat.off <- MkOffCat(LL.off[i,][these, drop=FALSE],  T.off=3)
	 ...
  }}
}

}
\seealso{
\code{\link{MkIDcat}, {MatchPedField}}
}
