% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offspring.R
\name{parDFtoM}
\alias{parDFtoM}
\title{turn long-format dataframe into a matrix}
\usage{
parDFtoM(df, expl.var = "prob")
}
\arguments{
\item{df}{dataframe subset, with at least columns 'id' and 'par.field'}

\item{expl.var}{name of column to become matrix contents}
}
\value{
a matrix with 'id' in rows and 'par.field' in columns. If "INVIS" is
  not among 'par.field', it will be added.
}
\description{
to be used in llply etc.
}
\examples{
\dontrun{
candpar.L <- dlply(candpar.df, "par.ped", parDFtoM, expl.var="prob")
}
}
