% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offspring.R
\name{ParentOfAll}
\alias{ParentOfAll}
\title{Column products}
\usage{
ParentOfAll(M, nInvis = 10)
}
\arguments{
\item{M}{matrix, with 'id' in rows and 'par.field' in columns, for a single
par.ped}

\item{nInvis}{single number, estimated (average) number of potential
invisible candidate parents for each individual}
}
\value{
a numeric vector with log10-probabilities, of length equal to the
  number of columns in M
}
\description{
probabilities for each candidate parent (columns) to be the
  parent of all offspring (rows), given a matrix of per-offspring per-parent
  probabilities for a single par.ped.
}
\details{
The result for 'INVIS' (invisible candidate parent) includes a
  correction factor, as the invisible parent of offspring A may differ from
  the invisible parent of B, as a function of the estimated total number of
  invisible candidate parents (\code{nInvis}, \eqn{n_I}) and the number of
  siblings \eqn{n_O}:
  \deqn{P(\text{INVIS}) = P'(\text{INVIS}) \times
         \left(\frac{1}{n_I}\right)^{n_O} \times n_I}
}
\examples{
\dontrun{
  candpar.L <- dlply(candpar.df, "par.ped", parDFtoM, expl.var="prob")
 Candpar.L2 <- llply(candpar.L, ParentOfAll, nInvis[p])
}
}
