% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{ReadTable}
\alias{ReadTable}
\title{wrapper for read.table() with better defaults}
\usage{
ReadTable(file, sep = "\\t", header = TRUE, na.strings = c("", NA), ...)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{sep}{the field separator character.}

\item{header}{a logical value indicating whether the file contains the names
of the variables as its first line.}

\item{na.strings}{a character vector of strings which are to be interpreted
as NA values.}

\item{...}{further arguments to \code{\link{read.table}}}
}
\value{
A data frame (data.frame) containing a representation of the data in the file.
}
\description{
has the defaults \code{\link{read.table}} should have had.
}
\details{
stringsAsFactors=FALSE.
}
